Rename primary folder to key_switcheroo
